+++ mjs/index.js
@@ -65,3 +65,5 @@ const fadeOut = (el) => {
 // show and hidden elements
 const hide = (el) => (el.style.display = 'none');
 const show = (el) => (el.style.display = '');
+// exports
+export { $body, $selector, $selectorAll, $id, $date, $getJSON, $ajax, $createElement, fadeIn, fadeOut, hide, show };